service dir refactor
