+++ src/app/modules/login/login-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import {LoginComponent} from "./login/login.component";
-import {RouteService} from "../../shared/services/route/route.service";
+import {RouteService} from "../../services/route/route.service";
 import {SignupComponent} from "./signup/signup.component";
 
 const routes: Routes = [
